% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_similarity_matrix.R
\name{compute_similarity_matrix}
\alias{compute_similarity_matrix}
\title{Compute Similarity Matrix Between TCGA and Mouse Model Data}
\usage{
compute_similarity_matrix(
  expression_matrix,
  group_all,
  tcga_sample_size,
  tcga_normal_count,
  merge_clinic_CNV,
  cnv_genes,
  clinical_vars
)
}
\arguments{
\item{expression_matrix}{The Batch-corrected expression data with genes as row names from \code{prepare_PCA_data()}.}

\item{group_all}{The group information of TCGA and all mouse models from \code{prepare_PCA_data()}.}

\item{tcga_sample_size}{Integer. \strong{Number of TCGA samples} in the dataset.}

\item{tcga_normal_count}{Integer. \strong{Number of TCGA normal samples}.}

\item{merge_clinic_CNV}{Data frame. Processed TCGA \strong{clinical + CNV} data from \code{prepare_similarity_data()}.}

\item{cnv_genes}{Character vector. Gene names for CNV annotations (e.g., \code{c("TP53", "PTEN")}).}

\item{clinical_vars}{Character vector. \strong{User-defined clinical variables} to extract (e.g., \code{c("Hepatitis B", "Non-Alcoholic")}).}
}
\value{
A list containing:
\describe{
\item{similarity_matrix}{The full transcriptional similarity matrix before background removal.}
\item{dm_trom_del_backgroud_clinic}{The filtered similarity matrix merged with clinical data.}
}
}
\description{
\itemize{
\item \strong{Formats} the input expression matrix by setting gene IDs as the first column.
\item Extracts \strong{TCGA and mouse model expression data} from a preprocessed matrix.
\item Computes a \strong{similarity matrix} using \code{ws.trom()}, comparing TCGA with mouse models.
\item Filters similarity data \strong{to retain only relevant comparisons} (tumor vs. tumor, normal vs. normal).
\item \strong{Incorporates TCGA clinical data} by merging with \code{merge_clinic_CNV}.
}
}
\details{
This function calculates a \strong{transcriptional similarity matrix} between \strong{TCGA samples} and \strong{mouse models}.
}
\examples{
\dontrun{
similarity_results <- compute_similarity_matrix(
  expression_matrix = pca_data$corrected_data,
  group_all = pca_data$group_all,
  tcga_sample_size = 424,
  tcga_normal_count = 50,
  merge_clinic_CNV = merge_clinic_CNV,
  cnv_genes = c("TP53", "PTEN"),
  clinical_vars = c("Hepatitis B", "Non-Alcoholic")
)
}

}
