% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_TROMscore_boxplot.R
\name{plot_TROMscore_boxplot}
\alias{plot_TROMscore_boxplot}
\title{Plot TROM Score Boxplot and Tileplot}
\usage{
plot_TROMscore_boxplot(
  merge_long,
  output_path = "./similarity/TROM_boxplot",
  combine_plots = TRUE
)
}
\arguments{
\item{merge_long}{Data frame. Output from \code{plot_TROMscore_heatmap()}, containing \strong{TROM scores}.}

\item{output_path}{Character. File path to save the figure (default: \code{"./similarity/TROM_boxplot"}).}

\item{combine_plots}{Logical. Whether to \strong{combine tileplot & boxplot} into one figure (default: \code{TRUE}).}
}
\value{
Saves the plot(s) and returns the ggplot object.
}
\description{
This function generates \strong{two subplots}:
\enumerate{
\item \strong{Boxplot} of transcriptional similarity scores (\verb{TROM score}).
\item \strong{Tileplot} showing the \strong{mean similarity scores per condition}.
}
}
\examples{
\dontrun{
plot_TROMscore_boxplot(
  merge_long = heatmap_results$merge_long,
  output_path = "./similarity/TROM_boxplot",
  combine_plots = TRUE
)
}

}
